+++ src/Misc/Player.cs
@@ -1516,5 +1516,24 @@ namespace mpvnet
                     track.Text += " " + value + ",";
             }
         }
+
+        private string[] _ProfileNames;
+
+        public string[] ProfileNames
+        {
+            get
+            {
+                if (_ProfileNames == null)
+                {
+                    string[] ignore = { "builtin-pseudo-gui", "encoding", "libmpv", "pseudo-gui", "default" };
+                    string profileList = Core.GetPropertyString("profile-list");
+                    var json = profileList.FromJson<List<Dictionary<string, object>>>();
+                    _ProfileNames = json.Select(i => i["name"].ToString())
+                                        .Where(i => !ignore.Contains(i)).ToArray();
+                }
+
+                return _ProfileNames;
+            }
+        }
     }
 }